+++ src/Database/Persist/Monad.hs
@@ -71,6 +71,7 @@ module Database.Persist.Monad (
 ) where
 
 import Control.Monad.Catch (MonadCatch, MonadMask, MonadThrow)
+import Control.Monad.Fix (MonadFix)
 import Control.Monad.IO.Class (MonadIO (..))
 import Control.Monad.IO.Unlift (MonadUnliftIO (..), wrappedWithRunInIO)
 import Control.Monad.Logger (MonadLogger)
@@ -108,7 +109,7 @@ import Database.Persist.Monad.SqlQueryRep
 newtype SqlTransaction m a = SqlTransaction
   { unSqlTransaction :: SqlPersistT m a
   }
-  deriving (Functor, Applicative, Monad, MonadRerunnableIO)
+  deriving (Functor, Applicative, Monad, MonadFix, MonadRerunnableIO)
 
 instance
   ( GHC.TypeError ( 'GHC.Text "Cannot run arbitrary IO actions within a transaction. If the IO action is rerunnable, use rerunnableIO")
@@ -194,6 +195,7 @@ newtype SqlQueryT m a = SqlQueryT
     ( Functor
     , Applicative
     , Monad
+    , MonadFix
     , MonadIO
     , MonadTrans
     , MonadResource