+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Add support for `persistent-2.14`
 * Add `Database.Persist.Sql.Shim` module that re-exports `Database.Persist.Sql` and `Database.Persist.Monad.Shim`
 * Add `MonadFix` instance to `SqlQueryT` and `SqlTransaction`
+* Add `getSqlBackendPool`
 
 # 0.4.0.0