+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Add `MonadFix` instance to `SqlQueryT` and `SqlTransaction`
 * Add `getSqlBackendPool`
 * Move `rerunnableLift` into `MonadRerunnableTrans` typeclass in `Control.Monad.Trans.Rerunnable`
+* Add `catchSqlTransaction`
 
 # 0.4.0.0