+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+# v0.5.0.0
+
 * Export `SqlQueryT` constructor ([#46](https://github.com/brandonchinn178/persistent-mtl/pull/46))
 * Drop support for GHC < 8.10
 * Drop support for `persistent` < 2.13